requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.40.61
botocore==1.40.61
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.20
dill==0.4.0
distro==1.9.0
docstring-parser==0.17.0
filelock==3.24.3
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.1
idna==3.11
ijson==3.4.0.post0
inspect-ai==0.3.180
jiter==0.13.0
jmespath==1.1.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
markdown-it-py==4.0.0
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
multiprocess==0.70.18
nest-asyncio2==1.7.2
numpy==2.4.2
openai==2.21.0
openrouter==0.6.0
packaging==26.0
pandas==3.0.1
pathlib-abc==0.5.2
pillow==12.1.1
pip==25.2
platformdirs==4.9.2
ply==3.11
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyyaml==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
s3fs==2025.9.0
s3transfer==0.14.0
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
textual==8.0.0
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==1.0.3
universal-pathlib==0.3.10
urllib3==2.6.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
